[libc] Split system/dirent.c
